+++ CHANGELOG.md
@@ -32,6 +32,7 @@ Types of changes
 - Add warning if a `<v1.0.0` model is loaded using `v1.0.0` or newer.
 - Propagate `SpanMarkerModel.from_pretrained` kwargs to the encoder its `AutoModel.from_pretrained`.
 - Ignore `UndefinedMetricWarning` when evaluation f1 is 0.
+- Improved model card generation.
 
 ### Fixed